+++ evaluations/interpolate.py
@@ -179,7 +179,7 @@ def main():
             sample = sample.squeeze(0).cpu().numpy()
 
             file_path = save_path / f"interpolate_{lamda}_{run_id}.npy"
-            np.save(str(file_path), sample)
+            np.save(str(file_path), sample.astype(np.uint16))
             print(f"Saved to {file_path}")
 
     payload = vars(parser_args)
